+++ commons/zenoh-codec/src/zenoh/reply.rs
@@ -52,7 +52,7 @@ where
 
         // Body
         if consolidation != &Consolidation::default() {
-            self.write(&mut *writer, *consolidation as u64)?;
+            self.write(&mut *writer, *consolidation)?;
         }
 
         // Extensions
